requirements.txt
anyio==3.6.2
astroid==3.1.0
attrs==23.1.0
bcrypt==4.1.2
cattrs==22.2.0
click==8.1.7
colorama==0.4.6
dill==0.3.6
dnspython==2.6.1
docstring-to-markdown==0.12
ecdsa==0.18.0
fastapi==0.95.1
h11==0.14.0
httptools==0.6.1
idna==3.4
isort==5.12.0
jedi==0.18.2
jedi-language-server==0.40.0
lazy-object-proxy==1.9.0
lsprotocol==2023.0.0a1
mccabe==0.7.0
nodeenv==1.8.0
parso==0.8.3
passlib==1.7.4
platformdirs==3.2.0
pyasn1==0.6.0
pycodestyle==2.11.1
pydantic==1.10.7
pydocstyle==6.3.0
pyflakes==3.2.0
pygls==1.0.1
pylama==8.4.1
pymongo==4.6.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
starlette==0.26.1
tomlkit==0.11.7
typeguard==2.13.3
typing_extensions==4.5.0
uvicorn==0.27.1
watchfiles==0.21.0
websockets==12.0
wrapt==1.15.0
